requirements.txt
asgiref==3.7.2
Django==5.0.2
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2024.1
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2024.1
